ml-evaluation/requirements.txt
# General
requests
python-dotenv

# Evaluation Suite
ragas
google-generativeai
litellm
matplotlib